meet_llm/requirements.txt
fastapi
uvicorn
redis
python-dotenv
mistralai
faster-whisper
sounddevice
soundfile
numpy
reportlab